Separate assertions and testing utilities
