+++ base/src/test/java/io/spine/protobuf/MessageFieldExceptionTest.java
@@ -29,7 +29,7 @@ import org.junit.jupiter.api.Test;
 
 import static io.spine.testing.Assertions.assertNpe;
 import static io.spine.testing.TestValues.newUuidValue;
-import static io.spine.testing.Tests.nullRef;
+import static io.spine.testing.TestValues.nullRef;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
