+++ base/src/test/java/io/spine/string/AbstractStringifierTest.java
@@ -26,8 +26,8 @@ import org.junit.jupiter.api.Test;
 import java.lang.reflect.Modifier;
 
 import static com.google.common.testing.SerializableTester.reserializeAndAssert;
+import static io.spine.testing.Assertions.assertHasPrivateParameterlessCtor;
 import static io.spine.testing.Assertions.assertIllegalArgument;
-import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertTrue;
