+++ tools/smoke-tests/entity-queries-tests/src/test/java/io/spine/query/given/EntityQueryBuilderTestEnv.java
@@ -35,7 +35,7 @@ import java.util.stream.IntStream;
 
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static com.google.common.truth.Truth.assertThat;
-import static io.spine.testing.Tests.nullRef;
+import static io.spine.testing.TestValues.nullRef;
 
 /**
  * Test environment and utilities for {@link io.spine.query.EntityQueryBuilderTest}.
